Update CamFragment.kt : 파일 디렉토리 이동
